This PR assigns "unique" rate-limit settings to many game and API
endpoints, making certain endpoints have stricter limits than others.
Most of these settings are stored in their own classes so multiple
endpoint classes could share them, and to organize them more clearly.
Not all rate-limit settings are in their own classes yet. The global
rate-limit is also slightly stricter for now.

Not all settings are properly tested yet. Some may, in reality, be too
strict or too loose; they are experimental.

+++ Refresh.Interfaces.APIv3/Endpoints/NotificationApiEndpoints.cs
@@ -1,9 +1,11 @@
 using AttribDoc.Attributes;
 using Bunkum.Core;
 using Bunkum.Core.Endpoints;
+using Bunkum.Core.RateLimit;
 using Bunkum.Protocols.Http;
 using MongoDB.Bson;
 using Refresh.Core.Authentication.Permission;
+using Refresh.Core.RateLimits.Users;
 using Refresh.Core.Types.Data;
 using Refresh.Database;
 using Refresh.Database.Models.Notifications;
@@ -20,6 +22,8 @@ public class NotificationApiEndpoints : EndpointGroup
 {
     [ApiV3Endpoint("notifications"), MinimumRole(GameUserRole.Restricted)]
     [DocUsesPageData, DocSummary("Gets a list of notifications stored for the user")]
+    [RateLimitSettings(NotificationsEndpointLimits.TimeoutDuration, NotificationsEndpointLimits.ApiRequestAmount, 
+                                NotificationsEndpointLimits.BlockDuration, NotificationsEndpointLimits.ApiRequestBucket)]
     public ApiListResponse<ApiGameNotificationResponse> GetNotifications(RequestContext context, GameUser user,
         GameDatabaseContext database, DataContext dataContext)
     {
@@ -32,6 +36,8 @@ public class NotificationApiEndpoints : EndpointGroup
     [DocSummary("Gets a specific notification for a user")]
     [DocError(typeof(ApiValidationError), ApiValidationError.ObjectIdParseErrorWhen)]
     [DocError(typeof(ApiNotFoundError), "The notification cannot be found")]
+    [RateLimitSettings(NotificationsEndpointLimits.TimeoutDuration, NotificationsEndpointLimits.ApiRequestAmount, 
+                                NotificationsEndpointLimits.BlockDuration, NotificationsEndpointLimits.ApiRequestBucket)]
     public ApiResponse<ApiGameNotificationResponse> GetNotificationByUuid(RequestContext context, GameUser user,
         GameDatabaseContext database,
         [DocSummary("The UUID of the notification")]
@@ -42,6 +48,7 @@ public class NotificationApiEndpoints : EndpointGroup
 
         GameNotification? notification = database.GetNotificationByUuid(user, objectId);
         if (notification == null) return ApiNotFoundError.Instance;
+        if (notification.UserId != user.UserId) return ApiNotFoundError.Instance; // 404 on purpose, notifications should remain private
         
         return ApiGameNotificationResponse.FromOld(notification, dataContext);
     }
@@ -50,6 +57,8 @@ public class NotificationApiEndpoints : EndpointGroup
     [DocSummary("Clears an individual notification for a user")]
     [DocError(typeof(ApiValidationError), ApiValidationError.ObjectIdParseErrorWhen)]
     [DocError(typeof(ApiNotFoundError), "The notification cannot be found")]
+    [RateLimitSettings(NotificationsEndpointLimits.TimeoutDuration, NotificationsEndpointLimits.ApiRequestAmount, 
+                                NotificationsEndpointLimits.BlockDuration, NotificationsEndpointLimits.ApiRequestBucket)]
     public ApiOkResponse ClearNotificationByUuid(RequestContext context, GameUser user, GameDatabaseContext database,
         [DocSummary("The UUID of the notification")] string uuid)
     {
@@ -58,6 +67,8 @@ public class NotificationApiEndpoints : EndpointGroup
         
         GameNotification? notification = database.GetNotificationByUuid(user, objectId);
         if (notification == null) return ApiNotFoundError.Instance;
+
+        if (notification.UserId != user.UserId) return ApiNotFoundError.Instance;
         
         database.DeleteNotification(notification);
         return new ApiOkResponse();
@@ -65,6 +76,8 @@ public class NotificationApiEndpoints : EndpointGroup
     
     [ApiV3Endpoint("notifications", HttpMethods.Delete), MinimumRole(GameUserRole.Restricted)]
     [DocSummary("Clears all notifications stored for the user")]
+    [RateLimitSettings(NotificationsEndpointLimits.TimeoutDuration, NotificationsEndpointLimits.ApiRequestAmount, 
+                                NotificationsEndpointLimits.BlockDuration, NotificationsEndpointLimits.ApiRequestBucket)]
     public ApiOkResponse ClearAllNotifications(RequestContext context, GameUser user, GameDatabaseContext database)
     {
         database.DeleteNotificationsByUser(user);
